apex: remove regex rule for classes
This rule overwrites the already existing behavior of ctags, which is to
classify Java classes as "c". Classifying Java classes as "classes"
leads to unexpected behavior during scoring in Zoekt.

I assume apex.ctags exists to reuse Java's rules for Apex and augment
them with Apex specific rules. If that is the case we should be fine
with using the default behavior for classes and just introducing the
rule for "properties."

I might not have all the context, so there might be reason why we have
to classify Java classes as "classes"?