enable markdown (#14)
In the past markdown caused issues. But I don't know what they where to be
honest, and I think universal ctags has improved since then. I want to
experiment before the next patch release for the impact on dotcom/etc. From
local experimentation it is quite nice in Zoekt for it to uprank headings in
markdown files.