Improve identification of definitions in Groovy (#18)
I think this should provide better results in symbol search, and hence in
'Go to definition' too.

One problem with the approach here is that we will potentially return results
from local variables in symbol search too, as we cannot detect whether we
have a global variable, or a field, or a local variable by just regex matching
on a single line. However, that seems like an OK compromise for now.