Return non-fatal errors instead of logging (#19)
Universal ctags distinguishes between two types of error:
* Fatal: user command is invalid or incorrectly formatted, I/O errors. This
indicates a big issue in how ctags is being called.
* Non-fatal: file parsing issues like invalid UTF-8, unsupported language, or
all symbols can't be parsed)

Currently, we only return fatal errors and just log non-fatal errors at "info".
Instead, we should return an error in both of these cases, and let the callers
decide how they want to handle it. For example, maybe we want to record an
error in observability.

This PR introduces a new struct in the API called `ParseError` which callers
can check to see if the error was fatal, and what file triggered the error.
It's an API break, but I think it's fine as this repo is intended only for
Sourcegraph use.